Add create stream endpoint at /streams.
Had to include only field on streamform to include foreignkeys in the
form, elsewise they will not be considered and the form will be valid
without them, which we don't want.

Also had to make sure the db session is removed at the end of a test.
We should probably move to flask-testing, which will take care of this
for us.

Added tests for the create endpoint. Documentation was weirdly enough
in place before the implementation...