Add log config.
Logging configuration is stored in separate yaml files, and might vary
in different environments, and is thus configurable through the
LOG_CONF_PATH configuration variable. We disallow starting the app
without this parameter set, to make sure it's detected early if this
happens in production. This is ignored though if either the DEBUG
or the TESTING config var is True.