Move stream create endpoint /streams -> /movies/<movie_id>/createStreams
This was done to avoid hard-coding client side, we're now including a
_links object in movies, pointing to interesting endpoints, for now
only createStream for that movie. This makes sure the client doesn't need
to know a movie's ID to be able to create streams, and can rather just
traverse links given in the API dynamically.

Updated tests to use the new endpoint, updated documentation