Fix branched migrations
Migrations for the json migration was generated on a branch that was not
rebased on master, which had had another migration applied as well. Thus
the alembic history branched, adding two heads on top of a previous
migration. This fixes it back to a linear history, making sure the
migrations can be applied cleanly.